refactor(types): remove custom workflow types, use local definitions in examples
- Delete packages/zatca-core/src/workflow-types.ts
- Remove OnboardResult, SingleReportResult, BatchReportResult, ClientInvoiceRecord from SDK exports
- Move type definitions to example files where they're actually used (server.ts, client.ts)
- These are application-level types, not core SDK types
- Both packages build and test successfully (217 + 103 tests passing)
